Add coding conventions, project overview, and suggested commands documentation
- Created `coding_conventions.md` to outline coding standards and TypeScript guidelines.
- Added `project_overview.md` detailing the purpose, tech stack, and features of the perf-audit-cli project.
- Introduced `suggested_commands.md` for recommended commands during development and CLI execution.
- Established `task_completion_checklist.md` to ensure all necessary commands are executed upon task completion.
- Developed `testing_guidelines.md` to provide testing framework details and best practices.

Refactor ESLint configuration to include unused imports plugin

- Updated `eslint.config.js` to integrate `eslint-plugin-unused-imports` for better management of unused variables and imports.
- Adjusted linting rules to enhance code quality checks.

Update package version and dependencies

- Changed package version to 0.0.3 in `package.json` and `package-lock.json`.
- Added `eslint-plugin-unused-imports` as a dependency.

Refactor CLI and command implementations

- Simplified CLI version retrieval by utilizing a utility function to read `package.json`.
- Refactored command files to use new command option interfaces and removed redundant code.

Create new type definitions for commands, database, and plugins

- Added new TypeScript interfaces for command options, database records, and plugin configurations to improve type safety and code clarity.

Enhance unit tests by removing unnecessary imports and improving error handling

- Cleaned up test files by removing unused imports and simplifying error handling in the dashboard tests.
