+++ package.json
@@ -35,8 +35,8 @@
     "start": "node dist/bin/cli.js",
     "test": "vitest --coverage",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint src/",
-    "lint:fix": "eslint src/ --fix",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
     "format": "dprint fmt",
     "format:check": "dprint check"
   },
@@ -53,6 +53,7 @@
     "@vitest/coverage-v8": "^3.2.4",
     "dprint": "^0.47.0",
     "eslint": "^9.0.0",
+    "eslint-plugin-unused-imports": "^4.2.0",
     "supertest": "^7.1.4",
     "typescript": "^5.0.0",
     "typescript-eslint": "^8.0.0",
