+++ src/commands/budget.ts
@@ -1,14 +1,11 @@
 import ora from 'ora';
 import { BundleAnalyzer } from '../core/bundle-analyzer.ts';
-import { AuditResult } from '../types/config.ts';
+import type { BudgetOptions } from '../types/commands.ts';
+import type { AuditResult } from '../types/config.ts';
 import { CIIntegration } from '../utils/ci-integration.ts';
 import { loadConfig } from '../utils/config.ts';
 import { Logger } from '../utils/logger.ts';
 import { ConsoleReporter } from '../utils/reporter.ts';
-
-interface BudgetOptions {
-  format: 'json' | 'console';
-}
 
 export async function budgetCommand(options: BudgetOptions): Promise<void> {
   // Only use spinner for non-JSON output to avoid interfering with JSON
