+++ src/types/bundle-analyzer.ts
@@ -1,0 +1,8 @@
+/**
+ * Bundle analyzer options interface
+ */
+export interface AnalyzeOptions {
+  outputPath: string;
+  gzip: boolean;
+  ignorePaths: string[];
+}
