+++ src/types/package.ts
@@ -1,0 +1,11 @@
+/**
+ * Package.json type definition
+ */
+export interface PackageJson {
+  name: string;
+  version: string;
+  description?: string;
+  author?: string;
+  license?: string;
+  [key: string]: unknown;
+}
