+++ test/unit/commands.test.ts
@@ -1,6 +1,5 @@
 import fs from 'fs';
-import path from 'path';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { initCommand } from '../../src/commands/init.ts';
 
 describe('Commands Unit Tests', () => {
