+++ test/unit/config.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { generateConfigFile, loadConfig } from '../../src/utils/config.ts';
 
 describe('Config', () => {
